+++ rust/src/transformations/make_stable_expr/expr_col/test.rs
@@ -1,6 +1,6 @@
 use crate::metrics::SymmetricDistance;
 use crate::transformations::StableExpr;
-use crate::transformations::make_stable_expr::test::get_test_data;
+use crate::transformations::make_stable_expr::test_helper::get_test_data;
 
 use super::*;
